--- lib/src/screens/post/viewmodel/hotpepperApiResponse.dart
@@ -1,1 +1,0 @@
-// TODO Implement this library.
